decentralized-voting-system/requirements.txt
python-dotenv==0.19.0
web3==5.24.0
flask==2.0.1
face_recognition==1.3.0
opencv-python==4.5.3.56
numpy==1.21.2
pytest==7.0.1
dlib==19.24.0
